requirements.txt
forexconnect==1.6.1   # Версия зависит от вашей текущей установки
mplfinance==0.12.9b7  # Последняя стабильная версия
python-dotenv==1.0.0  # Для работы с файлом .env
pandas==2.0.3         # Для работы с данными
